+++ HISTORY.md
@@ -2,6 +2,9 @@
 
 ## in development
 
+### 2019-01-02
+* Objects, Blocks, Threads, GUI, Store: added third color dimension and transparency to pen
+
 ### 2019-01-01
 * Morphic: added HSL color conversion support
 * Morphic: fixed glitch in WorldMorph >> getGlobalPixelColor